video: tegra: add T21x support
Clean up builds for tegra21 when display is enabled.
Add features table for t210.
Add help text to Kconfig options.
Fix CONFIG macros for building on T210
Fixes build issues

Bug 1476583

Reviewed-on: http://git-master/r/412321
GVS: Gerrit_Virtual_Submit

+++ arch/arm/mach-tegra/tegra12x_la.c
@@ -1101,8 +1101,8 @@ static void program_scaled_la(struct la_client_info *ci, int la)
 }
 static void program_la(struct la_client_info *ci, int la)
 {
-	unsigned long reg_read;
-	unsigned long reg_write;
+	u32 reg_read;
+	u32 reg_write;
 
 	BUG_ON(la > T12X_MC_LA_MAX_VALUE);
